Feat: 스터디 , 강의 상세 api 리팩토링 및 스터디 상태 변경 (모집완료/시작/종료) api 구현, 강의 상세페이지 api 연결 리팩토링 (#55)
* Refactor: 스터디 상세보기 api 관련 강의 정보 리턴하도록 수정

* Refactor: 강의 상세보기 api 리턴값에 관련 스터디 리스트 추가

* Refactor: 스터디 생성 api 리턴 값에 관련 강의 정보 추가

* Feat: 스터디 완료 api 구현

* Feat: 스터디 모집 완료 api 구현

* Feat: 스터디 시작 api 구현 완료

* Chore: lecture 스키마 수정

* Feat: Study 스키마 imageUrl required 제거, category required 추가

* Refactor: getLectureAction 수정

* Refactor: 강의 상세페이지 api 연결 리팩토링

+++ src/app/api/study/[id]/route.ts
@@ -1,4 +1,6 @@
 import connectDB from "@/lib/database/db";
+import { RecruitComment } from "@/schema/RecruitCommentSchema";
+import { Lecture } from "@/schema/lectureSchema";
 import { Study } from "@/schema/studySchema";
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
@@ -10,8 +12,15 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 
   try {
     const study = await Study.findById(studyId);
+    if (!study) {
+      return Response.json({ message: "스터디가 존재하지 않습니다." }, { status: 404 });
+    }
 
-    return Response.json(study);
+    const lectureId = study.lectureId;
+    const lectureData = await Lecture.findById(lectureId);
+    const { _id, category, online, platform, rating, title, instructor } = lectureData;
+    const lectureResult = { id: _id, category, online, platform, rating, title, instructor };
+    return Response.json({ study, lecture: lectureResult }, { status: 200 });
   } catch (error: any) {
     if (error.name === "CastError") {
       return Response.json({ message: "스터디가 존재하지 않습니다." }, { status: 404 });
@@ -28,6 +37,10 @@ export async function DELETE(request: Request, { params }: { params: { id: strin
   }
   connectDB();
   try {
+    const study = await Study.findById(studyId);
+    if (!study) {
+      return Response.json({ message: "스터디가 존재하지 않습니다." }, { status: 404 });
+    }
     await Study.findByIdAndDelete(studyId);
 
     return Response.json({ message: "스터디가 삭제되었습니다." }, { status: 200 });
